# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "FileMetadataDetail": {
        "properties": {
          "bitrate": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bitrate"
          },
          "codec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Codec"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "exif": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exif"
          },
          "extension": {
            "title": "Extension",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "image_height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Height"
          },
          "image_width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Width"
          },
          "md5": {
            "title": "Md5",
            "type": "string"
          },
          "metadata_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata Warning"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "pdf_author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Author"
          },
          "pdf_pages": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Pages"
          },
          "pdf_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Title"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "size_human": {
            "title": "Size Human",
            "type": "string"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "size_bytes",
          "size_human",
          "mime_type",
          "extension",
          "md5",
          "sha256",
          "uploaded_at"
        ],
        "title": "FileMetadataDetail",
        "type": "object"
      },
      "FileUploadResponse": {
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FileMetadataDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "size_human": {
            "title": "Size Human",
            "type": "string"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "key",
          "filename",
          "size_bytes",
          "size_human",
          "content_type",
          "uploaded_at"
        ],
        "title": "FileUploadResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperatorStat": {
        "description": "Per-operator effect within a run, parsed from the engine's own output.",
        "properties": {
          "filtered": {
            "title": "Filtered",
            "type": "integer"
          },
          "kept_ratio": {
            "title": "Kept Ratio",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "samples_in": {
            "title": "Samples In",
            "type": "integer"
          },
          "samples_out": {
            "title": "Samples Out",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "samples_in",
          "samples_out",
          "filtered",
          "kept_ratio"
        ],
        "title": "OperatorStat",
        "type": "object"
      },
      "PresignUploadRequest": {
        "description": "What the browser declares before uploading directly to B2.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "PresignUploadRequest",
        "type": "object"
      },
      "PresignUploadResponse": {
        "description": "A short-lived presigned PUT the browser uploads to, plus the exact\nheaders it must send. `Content-Length` and `content-type` are signed into\nthe URL, so B2 rejects a body of any other size or type.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "key",
          "url",
          "method",
          "content_type",
          "headers",
          "expires_in"
        ],
        "title": "PresignUploadResponse",
        "type": "object"
      },
...
      "RunRecord": {
        "description": "The full record of one curation pass.",
        "properties": {
          "dedup_ratio": {
            "default": 0.0,
            "title": "Dedup Ratio",
            "type": "number"
          },
          "device": {
            "default": "cpu",
            "title": "Device",
            "type": "string"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filtered": {
            "default": 0,
            "title": "Filtered",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kept_ratio": {
            "default": 0.0,
            "title": "Kept Ratio",
            "type": "number"
          },
          "modality": {
            "title": "Modality",
            "type": "string"
          },
          "operators": {
            "items": {
              "$ref": "#/components/schemas/OperatorStat"
            },
            "title": "Operators",
            "type": "array"
          },
          "recipe_id": {
            "title": "Recipe Id",
            "type": "string"
          },
          "recipe_name": {
            "title": "Recipe Name",
            "type": "string"
          },
          "refined_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refined Prefix"
          },
          "samples_in": {
            "default": 0,
            "title": "Samples In",
            "type": "integer"
          },
          "samples_out": {
            "default": 0,
            "title": "Samples Out",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "stats_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stats Key"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "recipe_id",
          "recipe_name",
          "modality",
          "status",
          "started_at"
        ],
        "title": "RunRecord",
        "type": "object"
      },
      "SeedResult": {
        "description": "Result of seeding a tiny synthetic demo corpus into ``raw/``.",
        "properties": {
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "modality": {
            "title": "Modality",
            "type": "string"
          },
          "prefix": {
            "title": "Prefix",
            "type": "string"
          }
        },
        "required": [
          "created",
          "prefix",
          "modality"
        ],
        "title": "SeedResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyUploadRequest": {
        "description": "Sent after the direct PUT so the API can inspect the stored object.",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          }
        },
        "required": [
          "key"
        ],
        "title": "VerifyUploadRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Local control-plane API for Data-Juicer Multimodal Curation: author and run Data-Juicer recipes that clean image-text, video, audio, and text corpora stored in Backblaze B2, writing refined shards and per-operator stats back to B2. This contract documents the app's local API, not a hosted public endpoint.",
    "title": "Data-Juicer Multimodal Curation API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/datasets/seed": {
      "post": {
        "description": "Generate a tiny synthetic demo corpus under ``raw/`` (keyless, no download).",
        "operationId": "seed_dataset_endpoint_datasets_seed_post",
        "parameters": [
          {
            "in": "query",
            "name": "modality",
            "required": false,
            "schema": {
              "default": "image-text",
              "title": "Modality",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Seed Dataset Endpoint",
        "tags": [
          "curation"
        ]
      }
    },
...
    "/runs/{run_id}": {
      "get": {
        "operationId": "get_run_endpoint_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Endpoint",
        "tags": [
          "runs"
        ]
      }
    },
    "/upload/presign": {
      "post": {
        "description": "Validate a declared upload and hand back a presigned PUT.\n\nThe browser uploads the bytes straight to B2 with the returned URL, so they\nnever traverse this Function - that is what lifts Vercel's ~4.5 MB payload\nceiling. Size and content-type are signed into the URL (see the service).",
        "operationId": "presign_upload_upload_presign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Presign Upload",
        "tags": [
          "upload"
        ]
      }
    },
    "/upload/verify": {
      "post": {
        "description": "Confirm an object just uploaded directly to B2 is valid and visible.",
        "operationId": "verify_upload_route_upload_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Upload Route",
        "tags": [
          "upload"
        ]
      }
    }
  }
}